# herald-server/src/http/events.rs
use std::sync::Arc;

use axum::extract::{Extension, Path, Query, State};
use axum::http::StatusCode;
use axum::response::IntoResponse;
use axum::Json;
use serde::Deserialize;

use herald_core::event::{Event, EventId};
use herald_core::protocol::*;

use crate::http::validation;
use crate::http::TenantId;
use crate::state::AppState;
use crate::store;
use crate::ws::connection::now_millis;
use crate::ws::fanout::fanout_to_stream;

#[derive(Deserialize, utoipa::ToSchema)]
pub struct InjectEventRequest {
    pub sender: String,
    pub body: String,
    #[serde(default)]
    pub meta: Option<serde_json::Value>,
    #[serde(default)]
    pub exclude_connection: Option<u64>,
    #[serde(default)]
    pub parent_id: Option<String>,
}

#[derive(Deserialize, utoipa::IntoParams)]
pub struct ListEventsQuery {
    pub before: Option<u64>,
    pub after: Option<u64>,
    pub limit: Option<u32>,
    pub thread: Option<String>,
}

const MAX_LIMIT: u32 = 100;

#[utoipa::path(
    post, path = "/streams/{id}/events",
    tag = "events",
    params(("id" = String, Path, description = "Stream ID")),
    security(("basic_auth" = []), ("bearer_auth" = [])),
    request_body = InjectEventRequest,
    responses(
        (status = 201, description = "Event published", body = crate::http::openapi::EventPublishResponse),
        (status = 400, description = "Validation error", body = crate::http::openapi::ErrorResponse),
        (status = 404, description = "Stream not found", body = crate::http::openapi::ErrorResponse),
    ),
)]
pub async fn inject_event(
    State(state): State<Arc<AppState>>,
    Extension(tenant): Extension<TenantId>,
    Path(stream_id): Path<String>,
    Json(req): Json<InjectEventRequest>,
) -> impl IntoResponse {
    let tid = &tenant.0;

    if let Err(e) = validation::validate_id(&req.sender, "sender") {
        return (*e).into_response();
    }
    if let Err(e) = validation::validate_body(&req.body) {
        return (*e).into_response();
    }
    if let Err(e) = validation::validate_meta(&req.meta) {
        return (*e).into_response();
    }
    if let Err(e) = validation::validate_attachments(&req.meta) {
        return (*e).into_response();
    }

    if !state.streams.stream_exists(tid, &stream_id) {
        return (
            StatusCode::NOT_FOUND,
            Json(serde_json::json!({"error": "stream not found"})),
        )
            .into_response();
    }

    if state.streams.is_archived(tid, &stream_id) {
        return (
            StatusCode::BAD_REQUEST,
            Json(serde_json::json!({"error": "stream is archived"})),
        )
            .into_response();
    }

    // See ws/handler.rs handle_publish for why seq gaps on storage failure are accepted.
    let seq = match state.allocate_seq(tid, &stream_id).await {
        Ok(s) => s,
        Err(e) => {
            tracing::error!(tenant = tid, stream = %stream_id, "failed to allocate sequence: {e}");
            return (
                StatusCode::INTERNAL_SERVER_ERROR,
                Json(serde_json::json!({"error": "internal error"})),
            )
                .into_response();
        }
    };
    let now = now_millis();
    let event_id = uuid::Uuid::new_v4().to_string();

    let event = Event {
        id: EventId(event_id.clone()),
        stream_id: stream_id.clone(),
        seq,
        sender: req.sender.clone(),
        body: req.body.clone(),
        meta: req.meta.clone(),
        parent_id: req.parent_id.clone(),
        edited_at: None,
        sent_at: now,
    };

    // Record in local writes set BEFORE store write — the store subscription
    // fires synchronously during the write, so the backplane consumer on this
    // instance must see the key in the set before it processes the event.
    if state.config.cluster.enabled {
        let event_key = format!("{tid}/{stream_id}/{seq:020}").into_bytes();
        state.record_local_write(event_key);
    }

    let ttl = state.event_ttl_ms(tid);
    if let Err(e) = store::events::insert(&*state.db, tid, &event, now + ttl).await {
        tracing::error!(tenant = tid, stream = %stream_id, "failed to insert event: {e}");
        return (
            StatusCode::INTERNAL_SERVER_ERROR,
            Json(serde_json::json!({"error": "internal error"})),
        )
            .into_response();
    }

    state.increment_tenant_events(tid);

    let new_event = ServerMessage::EventNew {
        payload: EventNewPayload {
            stream: stream_id.clone(),
            id: event_id.clone(),
            seq,
            sender: req.sender.clone(),
            body: req.body.clone(),
            meta: req.meta.clone(),
            parent_id: req.parent_id.clone(),
            sent_at: now,
        },
    };
    let exclude = req
        .exclude_connection
        .map(crate::registry::connection::ConnId);
    fanout_to_stream(
        &state,
        tid,
        &stream_id,
        &new_event,
        exclude,
        Some(&req.sender),
    )
    .await;

    // Cache channel: update last event for new subscribers
    state.streams.set_last_event(tid, &stream_id, new_event);

    state.fire_webhook(
        tid,
        crate::webhook::WebhookEvent {
            event: "event.new".to_string(),
            stream: stream_id.clone(),
            id: Some(event_id.clone()),
            seq: Some(seq),
            sender: Some(req.sender),
            body: Some(req.body),
            meta: req.meta,
            sent_at: Some(now),
            user_id: None,
            role: None,
        },
    );

    (
        StatusCode::CREATED,
        Json(serde_json::json!({"id": event_id, "seq": seq, "sent_at": now})),
    )
        .into_response()
}

#[utoipa::path(
    get, path = "/streams/{id}/events",
    tag = "events",
    params(("id" = String, Path, description = "Stream ID"), ListEventsQuery),
    security(("basic_auth" = []), ("bearer_auth" = [])),
    responses(
        (status = 200, description = "List of events", body = crate::http::openapi::EventListResponse),
    ),
)]
pub async fn list_events(
    State(state): State<Arc<AppState>>,
    Extension(tenant): Extension<TenantId>,
    Path(stream_id): Path<String>,
    Query(q): Query<ListEventsQuery>,
) -> impl IntoResponse {
    let tid = &tenant.0;
    let limit = q.limit.unwrap_or(50).min(MAX_LIMIT);

    let events = if let Some(after) = q.after {
        store::events::list_after(&*state.db, tid, &stream_id, after, limit + 1)
            .await
            .unwrap_or_default()
    } else {
        let before = q.before.unwrap_or(i64::MAX as u64);
        store::events::list_before(&*state.db, tid, &stream_id, before, limit + 1)
            .await
            .unwrap_or_default()
    };

    let events = if let Some(ref thread_id) = q.thread {
        events
            .into_iter()
            .filter(|m| m.parent_id.as_deref() == Some(thread_id.as_str()))
            .collect()
    } else {
        events
    };

    let has_more = events.len() > limit as usize;
    let result: Vec<serde_json::Value> = events
        .into_iter()
        .take(limit as usize)
        .map(|m| {
            let mut obj = serde_json::json!({
                "id": m.id.0,
                "stream": m.stream_id,
                "seq": m.seq,
                "sender": m.sender,
                "body": m.body,
                "meta": m.meta,
                "sent_at": m.sent_at,
            });
            if let Some(ref parent_id) = m.parent_id {
                obj["parent_id"] = serde_json::json!(parent_id);
            }
            if let Some(edited_at) = m.edited_at {
                obj["edited_at"] = serde_json::json!(edited_at);
            }
            obj
        })
        .collect();

    Json(serde_json::json!({
        "events": result,
        "has_more": has_more,
    }))
    .into_response()
}

#[derive(Deserialize, utoipa::ToSchema)]
pub struct TriggerEphemeralRequest {
    pub event: String,
    #[serde(default)]
    pub data: Option<serde_json::Value>,
    #[serde(default)]
    pub exclude_connection: Option<u64>,
}

#[utoipa::path(
    post, path = "/streams/{id}/trigger",
    tag = "events",
    params(("id" = String, Path, description = "Stream ID")),
    security(("basic_auth" = []), ("bearer_auth" = [])),
    request_body = TriggerEphemeralRequest,
    responses(
        (status = 204, description = "Ephemeral event triggered"),
        (status = 404, description = "Stream not found", body = crate::http::openapi::ErrorResponse),
    ),
)]
pub async fn trigger_ephemeral(
    State(state): State<Arc<AppState>>,
    Extension(tenant): Extension<TenantId>,
    Path(stream_id): Path<String>,
    Json(req): Json<TriggerEphemeralRequest>,
) -> impl IntoResponse {
    let tid = &tenant.0;

    if !state.streams.stream_exists(tid, &stream_id) {
        return (
            StatusCode::NOT_FOUND,
            Json(serde_json::json!({"error": "stream not found"})),
        )
            .into_response();
    }

    let exclude = req
        .exclude_connection
        .map(crate::registry::connection::ConnId);

    let msg = ServerMessage::EventReceived {
        payload: EventReceivedPayload {
            stream: stream_id.clone(),
            event: req.event.clone(),
            sender: "_server".to_string(),
            data: req.data,
        },
    };
    fanout_to_stream(&state, tid, &stream_id, &msg, exclude, None).await;

    StatusCode::NO_CONTENT.into_response()
}

/// GDPR: Purge all events by a user in a stream, plus their cursors,
/// reactions, and blocks.
#[derive(Deserialize, utoipa::IntoParams)]
pub struct PurgeUserQuery {
    pub user_id: String,
}

#[utoipa::path(
    delete, path = "/streams/{id}/events",
    tag = "events",
    params(("id" = String, Path, description = "Stream ID"), PurgeUserQuery),
    security(("basic_auth" = []), ("bearer_auth" = [])),
    responses(
        (status = 200, description = "User events purged", body = crate::http::openapi::PurgeResponse),
        (status = 400, description = "Validation error", body = crate::http::openapi::ErrorResponse),
    ),
)]
pub async fn purge_user_events(
    State(state): State<Arc<AppState>>,
    Extension(tenant): Extension<TenantId>,
    Path(stream_id): Path<String>,
    Query(q): Query<PurgeUserQuery>,
) -> impl IntoResponse {
    let tid = &tenant.0;

    if let Err(e) = validation::validate_id(&q.user_id, "user_id") {
        return (*e).into_response();
    }

    match store::purge::purge_user_data(&*state.db, tid, &stream_id, &q.user_id).await {
        Ok(deleted) => {
            tracing::info!(
                tenant = tid,
                stream = %stream_id,
                user = %q.user_id,
                deleted,
                "GDPR: user data purged"
            );
            // Remove from in-memory member registry
            state.streams.remove_member(tid, &stream_id, &q.user_id);
            Json(serde_json::json!({"deleted": deleted})).into_response()
        }
        Err(e) => {
            tracing::error!(
                tenant = tid,
                stream = %stream_id,
                user = %q.user_id,
                "user purge failed: {e}"
            );
            (
                StatusCode::INTERNAL_SERVER_ERROR,
                Json(serde_json::json!({"error": "purge failed"})),
            )
                .into_response()
        }
    }
}