# openapi.yaml
openapi: 3.1.0
info:
  title: Herald API
  description: Persistent realtime event streams with built-in authorization.
  license:
    name: Proprietary
  version: 2.0.0
servers:
- url: /
  description: Local server
paths:
...
  /admin/events:
    get:
      tags:
      - admin
      summary: List recent admin events.
      operationId: list_events
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      - name: after_id
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
      responses:
        '200':
          description: Recent admin events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEventListResponse'
      security:
      - admin_auth: []
...
  /self/events:
    get:
      tags:
      - self-service
      summary: GET /self/events — tenant's recent admin events.
      operationId: list_events
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      - name: after_id
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
      responses:
        '200':
          description: Tenant admin events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEventListResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
...
  /streams/{id}/events:
    get:
      tags:
      - events
      operationId: list_events
      parameters:
      - name: id
        in: path
        description: Stream ID
        required: true
        schema:
          type: string
      - name: before
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
      - name: after
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
      - name: thread
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
    post:
      tags:
      - events
      operationId: inject_event
      parameters:
      - name: id
        in: path
        description: Stream ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InjectEventRequest'
        required: true
      responses:
        '201':
          description: Event published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPublishResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Stream not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
    delete:
      tags:
      - events
      operationId: purge_user_events
      parameters:
      - name: id
        in: path
        description: Stream ID
        required: true
        schema:
          type: string
      - name: user_id
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: User events purged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurgeResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
...
  /streams/{id}/members/{user_id}:
    delete:
      tags:
      - members
      operationId: remove_member
      parameters:
      - name: id
        in: path
        description: Stream ID
        required: true
        schema:
          type: string
      - name: user_id
        in: path
        description: User ID
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Member removed
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
    patch:
      tags:
      - members
      operationId: update_member
      parameters:
      - name: id
        in: path
        description: Stream ID
        required: true
        schema:
          type: string
      - name: user_id
        in: path
        description: User ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRequest'
        required: true
      responses:
        '200':
          description: Member role updated
        '400':
          description: Invalid role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
...
  /streams/{id}/trigger:
    post:
      tags:
      - events
      operationId: trigger_ephemeral
      parameters:
      - name: id
        in: path
        description: Stream ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerEphemeralRequest'
        required: true
      responses:
        '204':
          description: Ephemeral event triggered
        '404':
          description: Stream not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
components:
  schemas:
...
    AdminEventListResponse:
      type: object
      required:
      - events
      properties:
        events:
          type: array
          items: {}
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          description: Human-readable error message.
...
    EventListResponse:
      type: object
      required:
      - events
      - has_more
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventView'
        has_more:
          type: boolean
    EventPublishResponse:
      type: object
      required:
      - id
      - seq
      - sent_at
      properties:
        id:
          type: string
        sent_at:
          type: integer
          format: int64
        seq:
          type: integer
          format: int64
          minimum: 0
    EventView:
      type: object
      description: Event as returned by the list events API (field names differ from storage).
      required:
      - id
      - stream
      - seq
      - sender
      - body
      - sent_at
      properties:
        body:
          type: string
        edited_at:
          type:
          - integer
          - 'null'
          format: int64
        id:
          type: string
        meta: {}
        parent_id:
          type:
          - string
          - 'null'
        sender:
          type: string
        sent_at:
          type: integer
          format: int64
        seq:
          type: integer
          format: int64
          minimum: 0
        stream:
          type: string
...
    InjectEventRequest:
      type: object
      required:
      - sender
      - body
      properties:
        body:
          type: string
        exclude_connection:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
        meta: {}
        parent_id:
          type:
          - string
          - 'null'
        sender:
          type: string
...
    PurgeResponse:
      type: object
      required:
      - deleted
      properties:
        deleted:
          type: integer
          minimum: 0
...
    TriggerEphemeralRequest:
      type: object
      required:
      - event
      properties:
        data: {}
        event:
          type: string
        exclude_connection:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
    UpdateMemberRequest:
      type: object
      required:
      - role
      properties:
        role:
          type: string
...
  securitySchemes:
    admin_auth:
      type: http
      scheme: bearer
    basic_auth:
      type: http
      scheme: basic
    bearer_auth:
      type: http
      scheme: bearer
security:
- basic_auth: []
- bearer_auth: []
- admin_auth: []
...